ui structure refactor
