+++ docs/UI_readme.md
@@ -39,3 +39,7 @@ This guide explains how to set up and run the UI (`ui_start.py`) on a Raspberry 
 - If simulation page causes issues, comment out pyqtgraph.opengl usage in the code.
 
 This setup provides a functional UI for robot control. Report issues in the repo. 
+
+## Performance notes (Raspberry Pi)
+- The UI no longer imports OpenGL on startup. The `pyqtgraph.opengl` import was removed from `src/gradient_os/ui_start.py` to avoid slow and fragile OpenGL initialization on headless or low-power systems. If a future simulation view is added, load OpenGL lazily within that feature only.
+- Trajectory list refresh is now asynchronous. `RealControlPage.refresh_trajectory_list` sends `GET_TRAJECTORIES` and relies on the UI dispatcher to populate the list when the response arrives. This avoids blocking the UI thread for up to 3 seconds while waiting for UDP responses.
